# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tasks with JWT auth and spaces (personal + corporate). Register or login at /api/auth/*; use GET /api/auth/check to validate session. Send token as Authorization: Bearer \u0026lt;token\u0026gt; for /api/todos, /api/create, /api/spaces. Todos support space_id and assignee_id for corporate workspaces.",
        "title": "Task.grid API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/spaces": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List spaces for the current user (personal + corporate where member). Ensures personal space exists.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spaces"
                ],
                "summary": "List spaces",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.SpaceListItem"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new corporate space. Caller becomes owner and admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spaces"
                ],
                "summary": "Create corporate space",
                "parameters": [
                    {
                        "description": "name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateSpaceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SpaceListItem"
                        }
                    }
                }
            }
        },
        "/api/spaces/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get space details and members (for corporate). Caller must be a member.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spaces"
                ],
                "summary": "Get space by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SpaceDetail"
                        }
                    },
                    "403": {
                        "description": "Not a member"
                    },
                    "404": {
                        "description": "Space not found"
                    }
                }
            }
        },
...
        "/api/spaces/{id}/members": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a user to the space by email. Caller must be admin. User must already exist.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spaces"
                ],
                "summary": "Invite member to space",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "email, role",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.InviteMemberRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Not admin"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateSpaceRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "handlers.InviteMemberRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SpaceDetail": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.SpaceMember"
                    }
                },
                "my_role": {
                    "description": "current user's role in this space (admin, member)",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "handlers.SpaceListItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "handlers.SpaceMember": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}